feat: complete API standardization with enhanced services and UI
- Add MCP server dependencies and npm scripts for RestAssured integration
- Enhance all services with improved business logic and error handling
- Update view controllers with standardized response patterns
- Expand validation schemas and model relationships
- Improve UI templates with better error handling and user experience
- Add comprehensive test improvements and mock configurations
- Include RestAssured integration documentation and Maven examples
- Update database configuration and model associations

+++ src/models/GastoUnico.model.js
@@ -8,7 +8,7 @@ export function defineGastoUnico(sequelize) {
             autoIncrement: true,
         },
         descripcion: { type: DataTypes.STRING, allowNull: false },
-        monto: { type: DataTypes.FLOAT, allowNull: false },
+        monto: { type: DataTypes.DECIMAL(10, 2), allowNull: false },
         fecha: { type: DataTypes.DATEONLY, allowNull: false },
         categoria_gasto_id: { type: DataTypes.INTEGER, allowNull: false, references: { model: 'categorias_gasto', key: 'id' } },
         importancia_gasto_id: { type: DataTypes.INTEGER, allowNull: false, references: { model: 'importancias_gasto', key: 'id' } },
